pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "firfoto"
version = "0.1.0"
description = "Cross-platform photo culling assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Pillow>=10.4",
  "exifread>=3.0",
  "PySide6>=6.7",
  "rawpy>=0.24",
]

[project.scripts]
firfoto = "firfoto.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
